backend/requirements.txt
fastapi
uvicorn[standard]
requests
python-multipart
openai
supabase
python-dotenv
